+++ src/werkzeug/routing/converters.py
@@ -19,6 +19,7 @@ class BaseConverter:
 
     regex = "[^/]+"
     weight = 100
+    part_isolating = True
 
     def __init__(self, map: "Map", *args: t.Any, **kwargs: t.Any) -> None:
         self.map = map
@@ -50,6 +51,8 @@ class UnicodeConverter(BaseConverter):
     :param length: the exact length of the string.
     """
 
+    part_isolating = True
+
     def __init__(
         self,
         map: "Map",
@@ -80,6 +83,8 @@ class AnyConverter(BaseConverter):
                   arguments.
     """
 
+    part_isolating = True
+
     def __init__(self, map: "Map", *items: str) -> None:
         super().__init__(map)
         self.regex = f"(?:{'|'.join([re.escape(x) for x in items])})"
@@ -97,6 +102,7 @@ class PathConverter(BaseConverter):
 
     regex = "[^/].*?"
     weight = 200
+    part_isolating = False
 
 
 class NumberConverter(BaseConverter):
@@ -107,6 +113,7 @@ class NumberConverter(BaseConverter):
 
     weight = 50
     num_convert: t.Callable = int
+    part_isolating = True
 
     def __init__(
         self,
@@ -168,6 +175,7 @@ class IntegerConverter(NumberConverter):
     """
 
     regex = r"\d+"
+    part_isolating = True
 
 
 class FloatConverter(NumberConverter):
@@ -191,6 +199,7 @@ class FloatConverter(NumberConverter):
 
     regex = r"\d+\.\d+"
     num_convert = float
+    part_isolating = True
 
     def __init__(
         self,
@@ -216,6 +225,7 @@ class UUIDConverter(BaseConverter):
         r"[A-Fa-f0-9]{8}-[A-Fa-f0-9]{4}-"
         r"[A-Fa-f0-9]{4}-[A-Fa-f0-9]{4}-[A-Fa-f0-9]{12}"
     )
+    part_isolating = True
 
     def to_python(self, value: str) -> uuid.UUID:
         return uuid.UUID(value)